requirements/prod.txt
aiomysql==0.0.20
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
idna==2.8
peewee-async==0.6.3a0
peewee==3.12.0
pycparser==2.19
pyjwt==1.7.1
pymysql==0.9.2
python-dotenv==0.10.3
redis==3.3.11
requests==2.22.0
six==1.13.0
tornado==6.0.3
urllib3==1.25.7
werkzeug==0.16.0
wtforms-json==0.3.3
wtforms-tornado==0.0.2
wtforms==2.2.1
